requirements.txt
Flask==3.1.2
Werkzeug==3.1.3

Flask-MySQLdb==2.0.0
mysqlclient==2.2.7

python-dotenv==1.1.0

twilio==9.5.2

sib-api-v3-sdk==7.6.0

pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1

APScheduler==3.10.4

google-api-python-client==2.149.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

gunicorn==23.0.0